Updating eclipse files (#24)
* Imported/exported the eclipse files
My eclipse version: 2024-12.
Only removed a date string from the .importorder
* Removed 2 files which were not/hardly used
The c formatting requires C support being added. It was intended for OpenCL, but I'm not sure if it's used. We probably won't miss it.
The eclipse_java_templates can be imported in Java -> Editor -> Templates. It will add option to code complete the license. Effectively when you type copyright (or part of it and press space) it will replace it with the copyright statement. I don't think we'll miss it as the copyright is also set when new files are created.
* Slightly adjusting pom.xml
Got warning about encoding when building locally. And for some reason I had the URL change locally, might as well add it now...